+++ libs/balances-and-allowances/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [3.0.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v2.4.0...balances-and-allowances-v3.0.0) (2026-03-11)
+
+
+### ⚠ BREAKING CHANGES
+
+* Remove support for Lens network ([#7128](https://github.com/cowprotocol/cowswap/issues/7128))
+
+### ✨ Features
+
+* remove support for Lens network ([#7128](https://github.com/cowprotocol/cowswap/issues/7128)) ([ffdd36d](https://github.com/cowprotocol/cowswap/commit/ffdd36db765eb1feb7fc675347de40b1e6df2bdd))
+
+
+### 🔧 Miscellaneous
+
+* **hotfix:** 2026-03-10 ([8c831f7](https://github.com/cowprotocol/cowswap/commit/8c831f70685c058f460d7a0fabbf2941163b9fba))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 3.0.0
+    * @cowprotocol/common-hooks bumped to 3.0.0
+    * @cowprotocol/common-utils bumped to 3.0.0
+    * @cowprotocol/core bumped to 3.0.0
+    * @cowprotocol/multicall bumped to 3.0.0
+    * @cowprotocol/tokens bumped to 3.0.0
+    * @cowprotocol/types bumped to 4.0.0
+    * @cowprotocol/wallet bumped to 3.0.0
+    * @cowprotocol/wallet-provider bumped to 2.1.8
+
 ## [2.4.0](https://github.com/cowprotocol/cowswap/compare/balances-and-allowances-v2.3.2...balances-and-allowances-v2.4.0) (2026-03-10)